% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_names.R
\name{check_names}
\alias{check_names}
\title{Check if character vector contains variable names}
\usage{
check_names(df, var_names, boolean = T)
}
\arguments{
\item{df}{a dataframe.}

\item{var_names}{Chracter vector to be compared with the dataframe names.}

\item{boolean}{Boolean object used to define if the output is going to be a boolean object \code{TRUE} , or a string \code{FALSE}. Default: \code{TRUE}.}
}
\description{
Function used to check if a string, or a character vector contains variable names of a given dataframe.
}
\details{
Function used to check if a string, or a character vector contains variable names of a given dataframe. 
This functions is mainly used to errorproof other functions of this package,
}
\examples{
library(forestmangr)

check_names(iris, "Species")
check_names(iris, "Species", boolean = F )

check_names(iris, c("Especies", "Setal.Width") )
check_names(iris, c("Especies", "Setal.Width"), boolean = F)
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
