% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_dbh.R
\name{eq_dbh}
\alias{eq_dbh}
\title{Calculate the equivalent diameter}
\usage{
eq_dbh(df, dbh, sep = "+", dbh_name = "dbh", dec_comma = FALSE,
  remove_old = TRUE)
}
\arguments{
\item{df}{A dataframe.}

\item{dbh}{Quoted name of the diameter at breast height variable. Used to filter out trees with no diameter measerument.}

\item{sep}{Separator used in the dbh variable to separate the values. Can be either: \code{";"}, \code{","} \code{"-"} , \code{" "} or \code{"+"}. Default: \code{"+"}.}

\item{dbh_name}{Character object for the new variable name. Default: \code{FALSE}.}

\item{dec_comma}{If \code{TRUE}, the decimals will be considred as a comma, if \code{FALSE}, decimals will be condidered as dot. Default: \code{}FALSE.}

\item{remove_old}{If \code{TRUE}, the old dbh variable will me removed from the dataframe. Default: \code{TRUE}.}
}
\value{
A dataframe with an added equivalent diameter variable.
}
\description{
Calculate the equivalent diameter of a character variable with multiple diameter variables
stored, separated by a symbol.
}
\examples{

library(forestmangr)
data("exfm22")
data("exfm23")

exfm22 
# This data has a column with dbh values separated by a "+" sign, and
# a dot as decimal, so we can just calculate the equivalent diameter directly:
eq_dbh(df = exfm22, dbh = "dbh", sep = "+", dbh_name = "eq_dbh")

exfm23
# In this data, decimals are separated by a comma. So we use the argument dec_comma = TRUE:
eq_dbh(df = exfm23, dbh = "dbh", sep = "+", dbh_name = "eq_dbh", dec_comma = TRUE)

# It's possible to keep the old variable, using remove_old=FALSE:
eq_dbh(df = exfm23, dbh = "dbh", sep = "+", dbh_name = "eq_dbh", dec_comma = TRUE, remove_old=FALSE)

}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
