% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exfm1.R
\docType{data}
\name{exfm1}
\alias{exfm1}
\title{Stratified random inventory pilot data}
\format{
A data frame with 22 observations and 4 variables:
\describe{
  \item{STRATA}{stratum number}
  \item{STRATA_AREA}{area of each strata, in hectares}
  \item{PLOT_AREA}{area of plots, in square meters}
  \item{VWB}{volume with bark, in cubic meters}
  \item{VWB_m3ha}{volume with bark, in cubic meters per hectare}
}
}
\source{
Soares, C. P. B., Paula Neto, F. and Souza, A. L. (2012) Dendrometria e Inventario Florestal. 2nd ed. Vicosa: UFV.
}
\usage{
data(exfm1)
}
\description{
In this data, each observation is a plot.
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
\keyword{data}
