% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exfm10.R
\docType{data}
\name{exfm10}
\alias{exfm10}
\title{Inventory data of an eucalyptus forest Brazil with dominant height variable}
\format{
A data frame with 900 observations and 14 variables:
\describe{
  \item{MAP}{map numbers}
  \item{PROJECT}{project number}
  \item{STRATA}{stratum number}
  \item{GENCODE}{genetic code of plots}
  \item{STRATA_AREA}{area of each strata, in hectares}
  \item{PLANTING_DATE}{date of planting}
  \item{SPACING}{Spacing used in the plots, in meters}
  \item{PLOT}{plot number}
  \item{MEASUREMENT_DATE}{date of measurement}
  \item{PLOT_AREA}{area of plots, in square meters}
  \item{PIT}{pit number}
  \item{DBH}{diameter at breast height, in centimeters}
  \item{TH}{total height, in meters}
  \item{OBS}{quality of trees, N = normal tree, D = dominant tree, F = a failure, or dead tree}
  \item{DH}{dominant height, in meters}
}
}
\usage{
data(exfm10)
}
\description{
In this data, each observation is a tree.
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
\keyword{data}
