% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exfm11.R
\docType{data}
\name{exfm11}
\alias{exfm11}
\title{Observed hieght values of trees, and estimated values using 3 different models}
\format{A data frame with 199 observations and 6 variables:
\describe{
  \item{STRATA}{stratum number}
  \item{PLOT}{plot number}
  \item{TH}{total height, in meters}
  \item{TH_EST1}{total estimated height with model 1, in meters}
  \item{TH_EST2}{total estimated height with model 2, in meters}
  \item{TH_EST3}{total estimated height with model 3, in meters}
}}
\usage{
data(exfm11)
}
\description{
In this data, each observation is a tree.
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
\keyword{data}
