% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exfm13.R
\docType{data}
\name{exfm13}
\alias{exfm13}
\title{Experimental data on nitrogen effect in species fertilizing}
\format{A data frame with 36 observations and 7 variables:
\describe{
  \item{species}{species common name}
  \item{trat}{treatment number}
  \item{esp}{species number}
  \item{N}{quantity of nitrogen used, in grams}
  \item{N2}{squared quantity of nitrogen used, in squared grams}
  \item{block}{block number}
  \item{dbh}{diameter at breast height, in meters}
}}
\usage{
data(exfm13)
}
\description{
In this data, each observation is a tree.
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
\keyword{data}
