% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exfm14.R
\docType{data}
\name{exfm14}
\alias{exfm14}
\title{Inventory data of an eucalyptus forest in minas gerais, with age and dominant height variables}
\format{A dataframe with 859 observations and 4 variables:
\describe{
  \item{strata}{stratum number}
  \item{plot}{plot number}
  \item{age}{average age of plots, in months}
  \item{dh}{dominant height, in meters}
}}
\usage{
data(exfm14)
}
\description{
In this data, each observation is a plot
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
\keyword{data}
