% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exfm15.R
\docType{data}
\name{exfm15}
\alias{exfm15}
\title{Simplified Inventory data of an eucalyptus forest in Minas Gerais, Brazil}
\format{A data frame with 900 observations and 7 variables:
\describe{
  \item{STRATA}{stratum number}
  \item{STRATA_AREA}{area of each strata, in hectares}
  \item{PLOT}{plot number}
  \item{PLOT_AREA}{area of plots, in square meters}
  \item{DBH}{diameter at breast height, in meters}
  \item{TH}{total height, in meters}
  \item{OBS}{quality of trees, N = normal tree, D = dominant tree, F = a failure, or dead tree}
}}
\usage{
data(exfm15)
}
\description{
In this data, each observation is a tree.
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
\keyword{data}
