% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exfm16.R
\docType{data}
\name{exfm16}
\alias{exfm16}
\title{Inventory data of an eucalyptus forest in Brazil, with age variable}
\format{A data frame with 139 observations and 8 variables:
\describe{
  \item{strata}{stratum number}
  \item{plot}{plot number}
  \item{age}{average age of plots, in months}
  \item{DH}{dominant height, in meters}
  \item{N}{number of individuals}
  \item{V}{volume of trees, in cubic meters}
  \item{B}{basal area, in square meters}
}}
\usage{
data(exfm16)
}
\description{
In this data, each observation is a plot.
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
\keyword{data}
