% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exfm17.R
\docType{data}
\name{exfm17}
\alias{exfm17}
\title{Inventory data of an eucalyptus forest in minas gerais, with age and site variables}
\format{A dataframe with 139 observations and 8 variables:
\describe{
  \item{strata}{stratum number}
  \item{plot}{plot number}
  \item{age}{average age of plots, in months}
  \item{DH}{dominant height, in meters}
  \item{N}{number of individuals}
  \item{V}{volume of trees, in cubic meters}
  \item{B}{basal area, in square meters}
  \item{S}{site variable, in meters}   
}}
\usage{
data(exfm17)
}
\description{
In this data, each observation is a plot. The site was estimated using Schumacher's model.
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
\keyword{data}
