% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exfm18.R
\docType{data}
\name{exfm18}
\alias{exfm18}
\title{Experimental data on nitrogen effect in species fertilizing}
\format{A data frame with 877 observations and 6 variables:
\describe{
  \item{Plot}{plot number}
  \item{Species}{species scientific name}
  \item{Tree}{tree number}
  \item{Trunk}{trunk number}
  \item{cbh}{circumference at breast height, in meters}
  \item{dbh}{diameter at breast height, in meters}
}}
\usage{
data(exfm18)
}
\description{
In this data, each observation is a tree.
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
\keyword{data}
