% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exfm19.R
\docType{data}
\name{exfm19}
\alias{exfm19}
\title{Volume of felled trees, measured in the Smalian method}
\format{
A data frame with 16 observations and 3393 variables:
\describe{
  \item{STRATA}{stratum number}
  \item{TREE}{number of trees}
  \item{DBH}{diameter at breast height, in centimeters}
  \item{TH}{total height, in meters}
  \item{CSA}{cross section area with bark, in square meters}
  \item{VWB}{volume with bark, in cubic meters}
  \item{FFWB}{form factor for each section}
  \item{FFWB_mean}{average form vactor value}
}
}
\usage{
data(exfm19)
}
\description{
In this data, each observation is a section of a tree.
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
\keyword{data}
