% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exfm2.R
\docType{data}
\name{exfm2}
\alias{exfm2}
\title{Stratified random inventory definite data}
\format{A dataframe with 57 observations and 4 variables:
\describe{
  \item{STRATA}{stratum number}
  \item{STRATA_AREA}{area of each strata, in hectares}
  \item{PLOT_AREA}{area of plots, in square meters}
  \item{VWB}{volume with bark, in cubic meters}
}}
\source{
Soares, C. P. B., Paula Neto, F. and Souza, A. L. (2012) Dendrometria e Inventário Florestal. 2nd edn. Viçosa: UFV.
}
\usage{
data(exfm2)
}
\description{
In this data, each observation is a plot.
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
\keyword{data}
