% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exfm20.R
\docType{data}
\name{exfm20}
\alias{exfm20}
\title{Inventory data of a natural forest in Brazil}
\format{
A data frame with 12295 observations and 18 variables:
\describe{
  \item{cod}{area code}
  \item{transect}{plot number}
  \item{tree}{tree number}
  \item{common.name}{species common name}
  \item{scientific.name}{species scientific name}
  \item{family}{species family name}
  \item{dbh}{diameter at breast height, in centimeters}
  \item{canopy.pos}{canopy position}
  \item{light}{level of light received by the tree}
  \item{dead}{tells if the tree is dead or not}
  \item{Hcom}{commercial height, in meters}
  \item{Htot}{total height, in meters}
  \item{date}{date of measurement}
  \item{utm.east}{utm east position value}
  \item{utm.north}{utm north position value}
  \item{vol}{volume of trees, in cubic meters}
  \item{plot.area}{plot area, in square meters}
  \item{total.area}{total area, in hectares}
}
}
\usage{
data(exfm20)
}
\description{
In this data, each observation is a tree.
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
\keyword{data}
