% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exfm22.R
\docType{data}
\name{exfm22}
\alias{exfm22}
\title{Revenue data if an eucalyptus forest}
\format{A data frame with 8 observations and 3 variables:
\describe{
  \item{year}{Revenue year}
  \item{cost}{cost values for that year, in dollars}
  \item{revenue}{revenue values for that year, in dollars}
}}
\usage{
data(exfm22)
}
\description{
In this data, each observation is the year's revenue.
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
\keyword{data}
