% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exfm4.R
\docType{data}
\name{exfm4}
\alias{exfm4}
\title{Simple random inventory definite data}
\format{A dataframe with 25 observations and 3 variables:
\describe{
  \item{TOTAL_AREA}{total area, in hectares}
  \item{PLOT_AREA}{area of plots, in square meters}
  \item{VWB}{volume with bark, in cubic meters}
}}
\source{
Soares, C. P. B., Paula Neto, F. and Souza, A. L. (2012) Dendrometria e Inventário Florestal. 2nd edn. Viçosa: UFV.
}
\usage{
data(exfm4)
}
\description{
In this data, each observation is a plot.
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
\keyword{data}
