% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exfm7.R
\docType{data}
\name{exfm7}
\alias{exfm7}
\title{Data of felled trees sections, measured in the Smalian method}
\format{A dataframe with 11 observations and 3393 variables:
\describe{
  \item{MAP}{map numbers}
  \item{PROJECT}{name of project}
  \item{SPACING}{Spacing used in the plots, in meters}
  \item{STRATA}{stratum number}
  \item{GENCODE}{genetic code of plots}
  \item{TREE}{number of trees}
  \item{DBH}{diameter at breast height, in meters}
  \item{TH}{total height, in meters}
  \item{hi}{height of sections, in meters}
  \item{di_wb}{diameter of sections with bark, in centimeters}
  \item{bark_t}{bark of thickness, in millimiters}
}}
\usage{
data(exfm7)
}
\description{
In this data, each observation is a section of a tree.
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
\keyword{data}
