% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exfm8.R
\docType{data}
\name{exfm8}
\alias{exfm8}
\title{Data of felled trees sections, measured in the Huber method}
\format{A data frame with 10 observations and 596 variables:
\describe{
  \item{CLONE}{Clone number}
  \item{STRATA}{stratum number}
  \item{TREE}{number of trees}
  \item{LOG}{log number}
  \item{DBH}{diameter at breast height, in meters}
  \item{TH}{total height, in meters}
  \item{hi}{height of sections, in meters}
  \item{di_wb}{diameter of sections with bark, in centimeters}
  \item{bark_t}{bark of thickness, in millimeters}
  \item{sec_length}{section length, in meters}
}}
\usage{
data(exfm8)
}
\description{
In this data, each observation is a section of a tree.
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
\keyword{data}
