% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FdeGraybill.R
\name{FdeGraybill}
\alias{FdeGraybill}
\title{Teste F de Graybill}
\usage{
FdeGraybill(df, Y1, Yj, output = 1, signif = 0.05)
}
\arguments{
\item{df}{Data frame a ser utilizado.}

\item{Y1}{Nome daVariavel padrao entre aspas.}

\item{Yj}{Nome Variavel proposta entre aspas.}

\item{output}{Tipo de saida simples (1), com algumas informacoes sobre as variaveis (2) ou completa (3). Padrao: \code{1}.}

\item{signif}{Valor da significancia a ser utilizada do teste. Padrao: \code{0.05}.}
}
\value{
Dataframe; suas dimensoes variam de acordo com o parametro \code{output}.
}
\description{
Teste de hipotese conforme descrito por GRAYBILL (1976).
}
\details{
Teste utilizado para comparar duas variaveis, geralmente um metodo proposto e um padrao.
Este teste é muito utilizado, principalmente na área florestal, sendo muitas vezes mais indicado do que os testes de média. Isto porque o teste <i>F</i> de Graybill considera todos os dados, e não só a média, que pode não ser representativa dos dados, principalmente com a presença de outliers. 

A validacao da predicao constitui-se em ajustar um modelo linear de 1º
grau dos valores preditos em funcao dos valores observados.

A significancia da regressao e avaliada aplicando-se o teste F para as
estimativas dos parametros, conforme metodologia descrita por GRAYBILL (1976).
}
\examples{
library(forestr)
data("exfm11")

# O data frame exfm11 possui estimativas de altura para uma determinada floresta.
# Pode-se determinar se o dado estimado e estatisticamente igual ao observado utilizando 
# o teste F de Graybill:

FdeGraybill( exfm11,"HT",  "HT_EST1")

# Pode-se alterar a significancia do teste utilizando o argumento signif:
FdeGraybill( exfm11,"HT",  "HT_EST1" , signif = 0.01)

# Para saidas mais detalhadas do teste, pode-se utilizar o argumento output:
FdeGraybill( exfm11,"HT",  "HT_EST2" , output=2)
FdeGraybill( exfm11,"HT",  "HT_EST2" , output=3)

}
\references{
CAMPOS, J. C. C.; LEITE, H. G. Mensuração florestal: perguntas e respostas. 3ª. ed. Viçosa: Editora UFV, 2013. 605 p.

GRAYBILL, F. A. Theory and application of the linear model. Massa chusetts: Ouxburg Press, p. 704, 1976.

LEITE, H. G.; OLIVEIRA, F. H. T. Statistical procedure to test identity between analytical methods. Communications in Soil Science Plant Analysis, v.33, n.7/8, p.1105-1118, 2002.
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
