% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huberwob.R
\name{huberwob}
\alias{huberwob}
\title{Calculate the volume without bark of trees using the Huber method}
\usage{
huberwob(df, di, section_length, bt, tree, .groups = NULL,
  di_mm_to_cm = FALSE, bt_mm_to_cm = FALSE)
}
\arguments{
\item{df}{A dataframe.}

\item{di}{Quoted name of the section diameter  variable, in centimeters.}

\item{section_length}{Quoted name of the section length variable, in meters}

\item{bt}{Quoted name of the bark thickness variable, in centimeters.}

\item{tree}{Quoted name of the tree variable. used to differentiate the trees' sections. If this argument is \code{NULL}, the defined groups in the dataframe will be used. Default: \code{NULL}.}

\item{.groups}{Optional argument. Quoted name(s) of additional grouping variables that can be added to differenciate subdivisions of the data. 
If this argument is \code{NULL}, the defined groups in the dataframe will be used. Default: \code{NULL}.}

\item{di_mm_to_cm}{Boolean argument that, if \code{TRUE}, converts the di argument from milimiters to centimeters. Default: \code{FALSE}.}
}
\value{
Dataframe with volume values by section.
}
\description{
Function used to calculate the volume without bark of trees using the Huber method.
This function has integration without dplyr, so it can be used inside a pipe, along with the
\code{group_by} function.
}
\examples{
library(forestmangr)
data("exfm8")

head(exfm8)

# Calculate the volume without bark using Huber's method:
huberwob(exfm8,"di_wb", "sec_length", "bark_t", "TREE")

# Using pipes:
library(dplyr)

 exfm8 \%>\% 
 group_by(TREE) \%>\% 
 huberwob("di_wb", "sec_length", "bark_t")

}
\references{
Campos, J. C. C. and Leite, H. G. (2017) Mensuração Florestal: Perguntas e Respostas. 5a. Viçosa: UFV.
}
\seealso{
Complementary functions:
  \code{\link{huberwb}}, For calculation of volume with bark using Huber's method,
  \code{\link{smalianwb}}, for calculation of volume with bark using Smalian's method,
  \code{\link{smalianwob}}, for calculation of volume without bark Smalian's method.
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
