% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ident_model.R
\name{ident_model}
\alias{ident_model}
\title{Teste de Identidade de Modelo}
\usage{
ident_model(df, factor, modelo_reduzido, filtrar = NULL,
  output = "table", grey_scale = F, signif = 0.05)
}
\arguments{
\item{df}{Data frame a ser utilizado.}

\item{factor}{Nome entre aspas da variavel classificatória utilizada no teste.}

\item{modelo_reduzido}{Modelo que sera utilizado no teste. Deve ser composto por nomes das variaveis, separando as variaveis dependendes e independentes com '~'.}

\item{filtrar}{Caso seja diferente de nulo, a funcao ira filtrar o dado com base na variavel factor fornecida, e os niveis inseridos aqui. Padrao: \code{NULL}.}

\item{output}{Argumento que indica se a saida sera a tabela anova, a tabela anova e o gráfico, ou uma lista contendo vários objetos, como tabela com variáveis dummy e regressões, além do gráfico. Pode ter como entrada "table", "table_plot" ou "full". Padrao: \code{"table"}.}

\item{signif}{Valor da significancia a ser utilizada do teste. Padrao: \code{0.05}.}
}
\value{
Por padrao, um dataframe contendo as informacoes sobre o teste. Caso
\code{saida_full = TRUE}, a saida sera uma lista, contendo alem da tabela do teste,
 um grafico ggplot, a tabela dummies e o summary dos ajuestes dos modelos reduzido e completo.
}
\description{
Funcao para se realizar o Teste de Identidade de Modelo Conforme descrito por Regazzi (1999)
}
\examples{
# Deseja-se saber se o comportamento do diametro e semelhante entre 3 especies.
# Sera utilizado o modelo quadratico.
# Primeiro visualiza-se os dados:
library(forestr)
data("ex13_mfr")
dados <- ex13_mfr
head(dados, 10)

# Agora basta rodar a função, especificando o nome da variavel que indica
# o projeto, e o modelo escolhido
ident_model(dados, "especie", dap ~ N + N2)

# Para se ter a saida completa com todas as tabelas e grafico de analise, utiliza-se:
ident_model(dados, "especie", dap ~  N + N2, output = "full")

# Caso o teste seja significativo e hajam mais de duas especies,
# e necessario fazer o teste com todas as combinacoes.
# Pode-se especificar em quais niveis se deseja fazer o teste
# com o argumento 'filtrar':

ident_model(dados, "especie", dap ~  N + N2, filtrar = c("PEQUI", "SUCUPIRA-PRETA"), output = "table_plot")
ident_model(dados, "especie", dap ~  N + N2, filtrar = c("PEQUI", "VINHATICO"), output = "table_plot")
ident_model(dados, "especie", dap ~  N + N2, filtrar = c("SUCUPIRA-PRETA", "PEQUI"), output = "table_plot")
ident_model(dados, "especie", dap ~  N + N2, filtrar = c("SUCUPIRA-PRETA", "VINHATICO"), output = "table_plot")

# Utilizando o argumento 'grey_scale' pode-se alterar a cor do grafico para tons de cinza:

ident_model(dados, "especie", dap ~  N + N2, output = "plot", grey_scale = T)

}
\references{
REGAZZI, A. J. Teste para verificar a identidade de modelos de regressão e a igualdade de parâmetros no caso de dados de delineamentos experimentais. Ceres, v. 46, n. 266, p. 383–409, 1999.
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}

Marcio leles Romarco de Oliveira \email{marcioromarco@gmail.com}
}
