% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ident_model_full.R
\name{ident_model_full}
\alias{ident_model_full}
\title{Identity of a Model complete Test}
\usage{
ident_model_full(df, factor, reduced_model, gray_scale = TRUE,
  signif = 0.05, font = "serif", output = "table_plot")
}
\arguments{
\item{df}{a data frame.}

\item{factor}{Quoted name of the factor variable used to differentiate the data projects in the test.}

\item{reduced_model}{Quoted or unquoted reduced model used in the test.  The variables mentioned in the model must exist in the provided data frame. X and Y sides of the model must be separated by "~".}

\item{gray_scale}{If \code{TRUE} a gray scale will be used in the plots. Default: \code{FALSE}.}

\item{signif}{Numeric value for the significance level used in the test. Default: \code{0.05}.}

\item{font}{font family used in the plots. Can be either \code{"serif"} for Times New Roman or \code{"sans"} for arial unicode MS. Default: \code{"serif"}.}

\item{output}{Defines the type of output. If \code{"table"} an anova table with the identity of model test is provided,
if \code{"plot"} a ggplot plot/object representing the final test is created, 
if \code{"plots"} in addition to the final test plot, the original test plot is also printed,
if \code{"table_plot"}, both anova table and plot are provided. Default: \code{"table_plot"}}
}
\value{
A list, or a ggplot object, varying according to the \code{output} argument.
}
\description{
Function for using the Identity of a Model test, as described by Regazzi (1999). 
Tests all combinations of factors and outputs the final result of the test.
}
\examples{
library(forestmangr)
data("exfm13")
exfm13

# The objective is to know if the diameter's behavior is similar among 3 species.
# For this we'll use a quadratic model. We'll use nitrogen (N) as our X variable.

ident_model_full(exfm13, "species", dbh ~ N + N2)

# After an initial test, differences are found. Since there are 3 different species,
# further testing is needed. After testing all combinations, we can see that
# SUcupira-Preta and Vinhatico are equal, and different from Pequi.
# Thus there should be 2 equations, one for pequi, and othe for Sucupira-Preta and Vinhatico.

}
\references{
Regazzi, A. J. (1999) Teste para verificar a identidade de modelos de regressao e a igualdade de parametros no caso de dados de delineamentos experimentais, Ceres, 46(266), pp. 383–409.
}
