% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_to_0.R
\name{na_to_0}
\alias{na_to_0}
\title{Convert NA to zero on numeric variables}
\usage{
na_to_0(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
a data frame
}
\description{
This function Convert NA to zero on numeric variables outside of mutate.
this is used due to changes on dplyr 1.1.0.
}
\examples{
library(forestmangr)
data("exfm15")
head(exfm15

# Turn NA values to zero only on numeric values
exfm15 \%>\% na_to_0()

}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
