% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliersiqr.R
\name{outliersiqr}
\alias{outliersiqr}
\title{Calculate interquartile range}
\usage{
outliersiqr(x)
}
\arguments{
\item{x}{a vector.}
}
\description{
Calculate interquartile range for a given vector
}
\examples{

library(forestmangr)
outliersiqr(iris$Sepal.Length)

}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
