% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuos.R
\name{residuos}
\alias{residuos}
\title{Graficos de Residuo}
\usage{
residuos(df, obs, ..., type = "scatterplot", nrow = NULL,
  ncol = NULL, lim_y = NULL, xlab = NULL, font = "serif",
  res_table = F)
}
\arguments{
\item{df}{Data frame a ser utilizado.}

\item{obs}{nome entre aspas da variavel observada.}

\item{...}{nome(s) entre aspas da(s) variavel(s) estimada(s). Multiplas variaveis devem ser separadas por virgula.}

\item{type}{tipo de grafico a ser gerado. Pode ser "scatterplot", "histogram", "histogram_curve", "versus". Padrao: \code{"scatterplot"}}

\item{lim_y}{Limite para o eixo y. Por padrao, este e definido automaticamente, mas pode ser ajustado caso necessario. Entra-se com um valor numero. Padrao: \code{NULL}.}

\item{res_table}{Caso verdadeiro, sera gerado um dataframe com os dados originais e seus respectivos erros. Padrao: \code{FALSE}.}
}
\value{
objeto ggplot, ou, caso \code{res_table = TRUE}, um dataframe contendo os dados originais e os residuos gerados.
}
\description{
Funcao para se criar graficos e tabelas de residuo a partir de dados observados e estimados.
}
\examples{
library(forestr)
data("ex11_mfr")

# Especificando-se apenas as variaveis observadas e estimadas,
# gera-se um grafico de dispersao:
residuos(ex11_mfr, "HT", "HT_EST1")
residuos(ex11_mfr, "HT", "HT_EST1", type = "scatterplot")

# Utilizando o argumento type, pode se criar outros tipos de grafico:
residuos(ex11_mfr, "HT", "HT_EST1", type = "histogram")
residuos(ex11_mfr, "HT", "HT_EST1", type = "histogram_curve")
residuos(ex11_mfr, "HT", "HT_EST1", type = "versus")

# Pode-se utilizar duas ou mais variáveis na comparaçao:
residuos(ex11_mfr, "HT", "HT_EST1", "HT_EST2", type = "scatterplot")
residuos(ex11_mfr, "HT", "HT_EST1", "HT_EST2", type = "histogram")
residuos(ex11_mfr, "HT", "HT_EST1", "HT_EST2", type = "histogram_curve")
residuos(ex11_mfr, "HT", "HT_EST1", "HT_EST2", type = "versus")

residuos(ex11_mfr, "HT", "HT_EST1", "HT_EST2", "HT_EST3")

#  Pode-se especificar os limites do eixo y com o argumento lim_y:
residuos(ex11_mfr, "HT", "HT_EST1", "HT_EST2", "HT_EST3", lim_y = 80)
residuos(ex11_mfr, "HT", "HT_EST1", "HT_EST2", "HT_EST3", lim_y = 40)

# E possivel gerar a tabela de resíduos com o argumento res_table:
head( residuos(ex11_mfr, "HT", "HT_EST1", "HT_EST2", res_table = T) )

}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
