% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_df.R
\name{round_df}
\alias{round_df}
\title{Round all numeric variables of a dataframe to a given digit}
\usage{
round_df(df, digits, rf = "round")
}
\arguments{
\item{df}{A dataframe.}

\item{digits}{Numeric vector for the desired number of digits.}

\item{rf}{Type of round to be used. It can either be \code{"ceiling"}, \code{"floor"}, \code{"trunc"}, \code{"signif"}, or \code{"round"}. Default \code("round").}
}
\value{
A dataframe, with all the numeric variables rounded up to the number given to \code{digits}.
}
\description{
This function allows the user to round all numeric values of a dataframe,
directly, even if the dataframe contains non-numeric variables
(which would throw an error in the \code{base::round} function).
}
\examples{
library(forestmangr)

# Round all numeric variables
round_df(iris) 

# Round all numeric variables using the floor function
round_df(iris, rf="floor") 

# Do not run
# trying this with the base::round function throws an error:

round(iris)

}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
