% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smaliancc.R
\name{smaliancc}
\alias{smaliancc}
\title{Volume com casca pelo metodo de Smalian}
\usage{
smaliancc(df, di, hi, .groups, di_mm_to_cm = FALSE, hi_cm_to_m = FALSE)
}
\arguments{
\item{df}{Data frame a ser utilizado.}

\item{di}{Nome entre aspas da variavel diametro da secao em centimetros.}

\item{hi}{Nome entre aspas da variavel altura da secao em metros.}

\item{.groups}{Nome(s) entre aspas da(s) variavel(s) classificatoria(s) utilizadas para identificar as arvores. Caso este argumento seja \code{NULL}, serao utilizados grupos ja definidos no dataframe. Padrao: \code{NULL}.}
}
\value{
Dataframe contendo os valores em nivel de secao.
}
\description{
Calculo do volume com casca em nivel de secao utilizando o metodo de Smalian.
}
\details{
Funcao utilizada para calcular o volume com casca pelo metodo de Smalian. A funcao possui
integracao com dplyr, podendo ser utilizada dentro de um pipe, em conjunto com a funcao
group_by.
}
\examples{
library(forestr)
data("ex7_mfr")

# Calcular o volume com casca pelo metodo de Smalian:
smaliancc(ex7_mfr,"di_cc", "hi", "ARVORE")

# ou, utilizando pipes:
 ex7_mfr \%>\% 
 group_by(ARVORE) \%>\% 
 smaliancc("di_cc", "hi")

}
\references{
CAMPOS, J. C. C.; LEITE, H. G. Mensuracao florestal: perguntas e respostas. 3a. ed. Vicosa: Editora UFV, 2013. 605 p.
}
\seealso{
Funcao complementar:
  \code{\link{smaliansc}}, para o calculo do volume sem casca.
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
