% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smaliansc.R
\name{smaliansc}
\alias{smaliansc}
\title{Volume sem casca pelo metodo de Smalian}
\usage{
smaliansc(df, di, hi, es, .groups, di_mm_to_cm = FALSE,
  hi_cm_to_m = FALSE, es_mm_to_cm = FALSE)
}
\arguments{
\item{df}{Data frame a ser utilizado.}

\item{di}{Nome entre aspas da variavel diametro da secao em centimetros.}

\item{hi}{Nome entre aspas da variavel altura da secao em metros.}

\item{es}{Nome entre aspas da variavel espessura da casca em milimetros.}

\item{.groups}{Nome(s) entre aspas da(s) variavel(s) classificatoria(s) utilizadas para identificar as arvores. Caso este argumento seja \code{NULL}, serao utilizados grupos ja definidos no dataframe. Padrao: \code{NULL}.}
}
\value{
Dataframe contendo os valores em nivel de secao.
}
\description{
Calculo do volume sem casca em nivel de secao utilizando o metodo de Smalian.
}
\details{
Funcao utilizada para calcular o volume sem casca pelo metodo de Smalian. A funcao possui
integracao com dplyr, podendo ser utilizada dentro de um pipe, em conjunto com a funcao
group_by.
}
\examples{
library(forestr)
data("ex7_mfr")

# Calcular o volume sem casca pelo metodo de Smalian:
smaliansc(ex7_mfr,"di_cc", "hi", "e_casca", "ARVORE")

# ou, utilizando pipes:
 ex7_mfr \%>\% 
 group_by(ARVORE) \%>\% 
 smaliansc("di_cc", "hi", "e_casca")

}
\references{
CAMPOS, J. C. C.; LEITE, H. G. Mensuracao florestal: perguntas e respostas. 3a. ed. Vicosa: Editora UFV, 2013. 605 p.
}
\seealso{
Funcao complementar:
  \code{\link{smaliancc}}, para o calculo do volume com casca.
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
