% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smalianwob.R
\name{smalianwob}
\alias{smalianwob}
\title{Calculate the volume without bark of trees using the Smalian method}
\usage{
smalianwob(df, di, hi, bt, .groups, di_mm_to_cm = FALSE,
  hi_cm_to_m = FALSE, bt_mm_to_cm = FALSE)
}
\arguments{
\item{df}{A dataframe.}

\item{di}{Quoted name of the section diameter  variable, in centimeters.}

\item{hi}{Quoted name of the section height  variable, in meters}

\item{bt}{Quoted name of the bark thickness variable, in centimeters.}

\item{.groups}{Quoted name of the factor variable(s) used to identify the trees. 
Aditional grouping variables can be added to differenciate subdivisions of the data. 
If this argument is \code{NULL}, the defined groups in the dataframe will be used. Default: \code{NULL}.}

\item{di_mm_to_cm}{Boolean argument that, if \code{TRUE}, converts the di argument from milimiters to centimeters. Default: \code{FALSE}.}

\item{hi_cm_to_m}{Boolean argument that, if \code{TRUE}, converts the hi argument from centimeters to meters. Default: \code{FALSE}.}

\item{nt_mm_to_cm}{Boolean argument that, if \code{TRUE}, converts the bt argument from milimiters to centimeters. Default: \code{FALSE}.}
}
\value{
Dataframe with volume values by section.
}
\description{
Function used to calculate the volume without bark of trees using the Smalian method.
This function has integration with dplyr, so it can be used inside a pipe, along with the
\code{group_by} function.
}
\examples{
library(forestmangr)
data("ex7_mfr")

head(ex7_mfr)

# Calculate the volume without bark using Smalian's method:
smalianwob(ex7_mfr,"di_wb", "hi", "bark_t", "TREE")

# Using pipes:
library(dplyr)

 ex7_mfr \%>\% 
 group_by(TREE) \%>\% 
 smalianwob("di_wb", "hi", "bark_t")

}
\references{
CAMPOS, J. C. C.; LEITE, H. G. Mensuracao florestal: perguntas e respostas. 3a. ed. Vicosa: Editora UFV, 2013. 605 p.
}
\seealso{
Complementary function:
  \code{\link{smalianwb}}, For calculation of the volume with bark.
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
