% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strs.R
\name{strs}
\alias{strs}
\title{Stratified Random Sampling}
\usage{
strs(df, Yi, plot_area, strata_area, .groups, age, alpha = 0.05,
  error = 10, dec_places = 4, pop = "inf", tidy = T)
}
\arguments{
\item{df}{a data frame.}

\item{Yi}{Quoted name of the volume variable, or other variable one desires to evaluate, in quotes.}

\item{plot_area}{Quoted name of the plot area variable, or a numeric vector with the plot area value. The plot area value must be in square meters.}

\item{strata_area}{Quoted name of the strata area variable, or a numeric vector with the plot strata values. If there are more than 1 area values, it's possible to use a vector with all area values, like so:\code{c(14.4, 16.4, 14.2)}. The strata area values must be in hectares.}

\item{.groups}{Quoted name of the subdivision variable(s), also known as strata. It's also possible to use more grouping variables.
In this case, the additional variables will be used to run multiple stratified surveys, one for each level of the additional groups,
while the last variable will be treated as the strata variable. In this case, one can use a character vector, like so: \code{c("MAP", "STRATA")}.}

\item{age}{Quoted name of the age variable. This is an optional parameter, that calculates the average age supplied.}

\item{alpha}{Numeric value for the significance value used in the t-student estimation. Default: \code{0.05}.}

\item{error}{Numeric value for the minimum admitted error value in the survey, in percentage. Default: \code{10}.}

\item{dec_places}{Numeric value for the number of decimal places to be used in the output tables. Default: \code{4}.}

\item{pop}{Character value for the type of population considered in the calculations. This can be either infinite (\code{"inf"}) or finite (\code{"fin"}). Default: \code{"inf"}.}

\item{tidy}{Boolean value that defines if the output tables should be tidied up or not. Default: \code{TRUE}.}
}
\value{
A list containing two dataframes, one with informations for each strata, and one with the stratified sampling results.
}
\description{
Function for processing forest inventory data using stratified random sampling.
}
\details{
This function allows the user to processes inventory data using stratified random sampling for n forest subdivisions (strata),
for finite or infinite populations. It's possible to run multiple sampling analysis using a factor variable indicated in the
\code{.groups}() parameter.
}
\examples{
library(forestmangr)
data("ex1_mfr")
data("ex2_mfr")
data("ex6_mfr")

# The objective is to sample an area, with an error of 5\%.
# First we run a pilot inventory, considering a 5\% error and a finite population:
strs(ex1_mfr, "VWB", "PLOT_AREA", "STRATA_AREA",  .groups = "STRATA", error = 5, pop = "fin")

# With these results, in order to meet the desired error of 5\%, we'll need to sample 59 more plots,
# 14 in stratum 1, 21 in stratum 2, and 24 in stratum 3.
# After sampling the necessary plots, we now run a definitive inventory,
# considering an 5\% error and a finite population:
strs(ex2_mfr, "VWB", "PLOT_AREA", "STRATA_AREA",  .groups = "STRATA", error = 5, pop = "fin")

# The desired error value was met.

# Area values can be numeric:
strs(ex2_mfr, "VWB", 1000, c(14.4, 16.4,14.2),  .groups = "STRATA", error = 5, pop = "fin")

# Optional variable age, and one stratified sampled inventory for each map:
strs(ex6_mfr, "VWB", "PLOT_AREA", "STRATA_AREA",  .groups = c("MAP","STRATA"), age = "AGE")

}
\references{
CAMPOS, J. C. C.; LEITE, H. G. Mensuracao florestal: perguntas e respostas. 3a. ed. Vicosa: Editora UFV, 2013. 605 p.

SOARES, C. P. B.; NETO, F. D. P.; SOUZA, A. L. D. Dendrometria e Inventario Florestal. 2a. ed. Vicosa: UFV, 2012. 272 p.
}
\seealso{
other sampling functions: 
  \code{\link{sprs}} for Simple Random Sampling, and
  \code{\link{ss_diffs}} for Systematic Sampling.
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
\keyword{Random}
\keyword{Sampling}
\keyword{Stratified}
