% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_summarise.R
\name{tree_summarise}
\alias{tree_summarise}
\title{Calculate the equivalent diameter of trees with more than one trunk}
\usage{
tree_summarise(df, dbh, tree, th = NULL, vwb = NULL, vwob = NULL,
  plot_area = NULL, total_area = NULL, .groups = NULL)
}
\arguments{
\item{df}{A dataframe.}

\item{dbh}{Quoted name of the diameter at breast height variable.}

\item{tree}{Quoted name of the tree variable. used to differentiate the trees' sections. If this argument is \code{NULL}, the defined groups in the dataframe will be used. Default: \code{NULL}.}

\item{th}{Optional argument. Quoted name of the total height variable, in meters. Default: \code{NULL}.}

\item{vwb}{Optional argument. Quoted name of the volume with bark varible, in cubic meters. Default: \code{NULL}.}

\item{vwob}{Optional argument. Quoted name of the volume without bark variable, in cubic meters. Default: \code{NULL}.}

\item{.groups}{Optional argument. Quoted name(s) of grouping variables that can be added to differenciate subdivisions of the data. Default: \code{NULL}.}
}
\value{
A dataframe with the the equivalent diameter calculated.
}
\description{
This function uses takes the square root of the diameters squared sum,
in order to estimate the equivalent diameter of trees. Other supplied variables
are summed up, or averaged, depending on the variable.
}
\examples{

library(forestmangr)
data("exfm18")
head(exfm18)

# Calculate the equivalent diameter of trees with more than one trunk:
tree_summarise(exfm18, "DBH",tree="Tree", .groups=c("Plot", "Species") )

}
\references{
Soares, C. P. B., Paula Neto, F. and Souza, A. L. (2012) Dendrometria e Inventário Florestal. 2nd edn. Viçosa: UFV.
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
