% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_stratum.R
\name{vertical_stratum}
\alias{vertical_stratum}
\title{Divide data into 3 vertical strata}
\usage{
vertical_stratum(df, th)
}
\arguments{
\item{df}{A data frame.}

\item{th}{Quoted name of the total height variable.}
}
\value{
a data frame.
}
\description{
Get the vertical strata of data based on the height variable.
The data will be divided into inferior, medium and superior strata.
}
\examples{
library(forestmangr)
data("exfm10")
exfm10

# To classify the data, supply the data frame and the height variable name:
vertical_stratum(exfm10, "TH" )

}
\references{
Souza, A. L. and Soares, C. P. B. (2013) Florestas Nativas: estrutura, dinamica e manejo. Vicosa: UFV.
}
\author{
Eric Bastos Gorgens \email{e.gorgens@gmail.com}
}
