% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_B2_clutter.R
\name{est_B2_clutter}
\alias{est_B2_clutter}
\title{Estimate future and present basal area values of Clutter's Growth and Yield Model}
\usage{
est_B2_clutter(df, age, basal_area, site, category, a0, a1,
  method = "average")
}
\arguments{
\item{df}{A dataframe.}

\item{age}{A numeric vector with the desired age range to be used in the estimation, or a Quoted name for the age variable.}

\item{basal_area}{Quoted name for the basal area variable.}

\item{site}{Quoted name for the average site variable.}

\item{category}{Quoted name for the category variable.}

\item{a0}{Numeric value for the a0 coefficient from Clutter's growth and yield model.}

\item{a1}{Numeric value for the a1 coefficient from Clutter's growth and yield model.}

\item{method}{Method used for estimating the present basal area of each class. It can either be the class' average basal area \code{"average"}, or an estimated value from a linear quadratic model of site as a function of basal area \code{"model"}. Default: \code{"average"}.}
}
\value{
A dataframe with the estimated values of basal area.
}
\description{
This function estimates the present the present value of basal area for each class
using either the class' mean, or a linear quadratic model, and then uses it's value to calculate
the basal area from clutter's growth and yield model.
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
