% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_table.R
\name{lm_table}
\alias{lm_table}
\title{Fit linear regressions by group, and get different output options.}
\usage{
lm_table(df, model, .groups = NA, output = "table", est.name = "est",
  keep_model = FALSE)
}
\arguments{
\item{df}{A dataframe.}

\item{model}{A linear regression model, with or without quotes. The variables mentioned in the model must exist in the provided dataframe. X and Y sides of the model must be separated by "~".}

\item{.groups}{Optional argument. Quoted name(s) of grouping variables used to fit multiple regressions, one for each level of the provided variable(s). Default \code{NA}.}

\item{output}{Selects different output options. Can be either \code{"table"}, \code{"merge"}, \code{"merge_est"} and \code{"nest"}. See details for explanations for each option. Default: \code{"table"}.}

\item{est.name}{Nmae of the estimated y value. Used only if \code{est.name = TRUE}. Padrao: \code{"est"}.}

\item{keep_model}{If \code{TRUE}, a column containg lm object(s) is kept in the output. Useful if the user desires to get more information on the regression.Default: \code{FALSE}.}
}
\value{
A dataframe. Different dataframe options are available using the output argument.
}
\description{
With this function it's possible to fit linear regressions by a grouping variable, and get a dataframe
with each column as a coefficient and quality of fit variables, and other output options. Works with dplyr grouping functions.
}
\details{
With this function there no more need to use the \code{do} function when fitting a linear regression in a pipe line.
It's also possible to easily make fit miltiple regressions, specifying a grouping variable.
In addition to that, the default output sets each coeffient as a column, making it easy to call coefficients by name or position
when estimating values. 

It's possible to use the \code{output} argument to get a merged table if \code{output="merge"}, that binds
the original dataframe and the fitted coefficients. 
If \code{output="merge_est"} we get a merged table aswell, but with y estimated using the coefficients. If the fit is made using groups, this is taken into account, i.e. the estimation is made by group.

If \code{output="nest"}, a dataframe with nested columns is provided. This can be used if the user desires to get a customized output.
}
\examples{
library(forestmangr)
library(dplyr)

data("exfm19")
head(exfm19)

# Fit Schumacher and Hall model for volume estimation, and get
# coefficient, R2 and error values:

lm_table(exfm19, log(VWB) ~  log(DBH) + log(TH))   

# Fit SH model by group:
lm_table(exfm19, log(VWB) ~  log(DBH) + log(TH), "STRATA")

# This can also be done using dplyr::group_by:
exfm19 \%>\% 
  group_by(STRATA) \%>\% 
  lm_table(log(VWB) ~  log(DBH) + log(TH) )
  
# It's possible to merge the original data with the table containg the coefficients
# using the output parameter:
lm_table(exfm19, log(VWB) ~  log(DBH) + log(TH), "STRATA", output = "merge")

# It's possible to merge the original data with the table,
# and get the estimated values for this model:
lm_table(exfm19, log(VWB) ~  log(DBH) + log(TH),"STRATA",
 output = "merge_est", est.name = "VWB_EST") \%>\% head(15)
   
# It's possible to further customize the output,
# unnesting the nested variables provided when output is defined as "nest":
lm_table(exfm19, log(VWB) ~  log(DBH) + log(TH),"STRATA", output = "nest")
 
 
# In the following example, the objective is to estimate non-measured height
# values in a forest inventory data.

# To do this, we'll fit a hipsometrical model. The non-measured trees will 
# be automatically ignored in this step. We'll define the argument output as
# "merge_est", so that we can get the estimated height values as a separate column.
# Then, we'll use mutate to create a new variable, that will contain measured
# height values, along with estimated ones. To do this we'll use ifelse, 
# and check for NAs inside the Height column. When it finds it, 
# it will basically fill them with estimated values.
library(forestmangr)
library(dplyr)

data("exfm15")
head(exfm15, 20)

ex_th_est <- exfm15 \%>\%  
  lm_table(log(TH) ~ inv(DBH),output = "merge_est" ) \%>\% 
  mutate( TH_EST = ifelse(is.na(TH), est, TH ) )
# Now we can see that the values were estimated successfully.
head(ex_th_est, 20)

}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
