% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npv_irr.R
\name{npv_irr}
\alias{npv_irr}
\title{Calculate Net Present Value and other economic variables}
\usage{
npv_irr(df, year, cost, revenue, rate, output = "vertical",
  big_mark = ",", dec_mark = ".")
}
\arguments{
\item{df}{A data frame.}

\item{year}{Quoted name of the year variable.}

\item{cost}{Quoted name of the costs variable.}

\item{revenue}{Quoted name of the revenue variable.}

\item{rate}{Numeric value of the yearly rate to be used.}

\item{output}{Selects different output options. It can be either \code{"horizontal"} for a data frame with a single observation and one column for each variable, or \code{vertical} for a two column data frame with one observation for each calculated variable. Default: \code{"vertical"}.}

\item{big_mark}{Selects thousands separator. Can be either \code{"."}, \code{" "} or \code{","}. Default: \code{,}.}

\item{dec_mark}{Selects decimal separator. Can be either \code{","} or \code{"."}. Default: \code{.}.}
}
\description{
Get the net present value, internal rate of return, and other 
economic variables, given cost and revenue values.
}
\examples{
library(forestmangr)

data(exfm22)

npv_irr(exfm22,"year","cost","revenue",rate=8.75)

npv_irr(exfm22,"year","cost","revenue",rate=8.75, "horizontal")

}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
