% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R, R/quasiquo.R
\name{\%>\%}
\alias{\%>\%}
\alias{:=}
\title{Pipe}
\description{
Pipe an object forward into a function or call expression. See \code{\link[magrittr]{magrittr}} for more details.

Search for quasiquotation for more details.
}
