% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_diffs.R
\name{ss_diffs}
\alias{ss_diffs}
\title{Systematic Sampling}
\usage{
ss_diffs(df, Yi, plot_area, total_area, age, .groups, alpha = 0.05,
  error = 10, dec_places = 4, tidy = T)
}
\arguments{
\item{df}{a data frame.}

\item{Yi}{Quoted name of the volume variable, or other variable one desires to evaluate, in quotes.}

\item{plot_area}{Quoted name of the plot area variable, or a numeric vector with the plot area value. The plot area value must be in square meters.}

\item{total_area}{Quoted name of the total area variable, or a numeric vector with the total area value.The total area value must be in hectares.}

\item{age}{Quoted name of the age variable. This is an optional parameter, that calculates the average age supplied.}

\item{.groups}{Quoted name of the subdivision variable(s), also known as strata. It's also possible to use more grouping variables.
In this case, the additional variables will be used to run multiple stratified surveys, one for each level of the additional groups,
while the last variable will be treated as the strata variable. In this case, one can use a character vector, like so: \code{c("MAP", "STRATA")}.}

\item{alpha}{Numeric value for the significance value used in the t-student estimation. Default: \code{0.05}.}

\item{error}{Numeric value for the minimum admitted error value in the survey, in percentage. Default: \code{10}.}

\item{dec_places}{Numeric value for the number of decimal places to be used in the output tables. Default: \code{4}.}

\item{tidy}{Boolean value that defines if the output tables should be tidied up or not. Default: \code{TRUE}.}
}
\value{
A dataframe with the sampling results
}
\description{
Function for processing forest inventory data using systematic sampling.
}
\details{
This function allows the user to processes inventory data using simple random sampling for finite or infinite populations.
It's possible to run multiple sampling analysis using a factor variable indicated in the \code{.groups}() parameter.
}
\examples{
library(forestmangr)
data("ex2_mfr")
data("ex5_mfr")

# We're trying to run a inventory for an area This data was colected systematically,
# but we'll try to run the data using simple random sampling, to show the difference between the two methods:
sprs(ex5_mfr,  "VWB", "PLOT_AREA", "TOTAL_AREA")

# We get a 22\% error value. Now, we run this same data
# considering the data as a systematic inventory, using the
# successive differences method:
ss_diffs(ex5_mfr,  "VWB", "PLOT_AREA", "TOTAL_AREA")

# The error was significantly lowered,

# Area Values can be numeric;
ss_diffs(ex5_mfr,  "VWB", 200, 18)

# Here we run a systematic sampling inventory for each forest subdivision, using the STRATA variable as a group variable:
ss_diffs(ex2_mfr,  "VWB", "PLOT_AREA", "STRATA_AREA",.groups = "STRATA")

}
\references{
CAMPOS, J. C. C.; LEITE, H. G. Mensuracao florestal: perguntas e respostas. 3a. ed. Vicosa: Editora UFV, 2013. 605 p.

SOARES, C. P. B.; NETO, F. D. P.; SOUZA, A. L. D. Dendrometria e Inventario Florestal. 2a. ed. Vicosa: UFV, 2012. 272 p.
}
\seealso{
other sampling functions: 
  \code{\link{sprs}} for Simple Random Sampling, and
  \code{\link{strs}} for stratified random sampling, and
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
\keyword{Sampling}
\keyword{Systematic}
