% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_summarise.R
\name{vol_summarise}
\alias{vol_summarise}
\title{Summarise volume of trees}
\usage{
vol_summarise(df, dbh, th, vwb, tree, .groups = NULL, vwob = NULL)
}
\arguments{
\item{df}{A dataframe.}

\item{dbh}{Quoted name of the diameter at breast hight variable, im cm.}

\item{th}{Quoted name of the total height variable, in meters.}

\item{vwb}{Quoted name of the volume with bark varible, in cubic meters.}

\item{tree}{Quoted name of the tree variable. used to differentiate the trees' sections. If this argument is \code{NULL}, the defined groups in the dataframe will be used. Default: \code{NULL}.}

\item{.groups}{Optional argument. Quoted name(s) of additional grouping variables that can be added to differenciate subdivisions of the data.}

\item{vwob}{Optional argument. Quoted name of the volume without bark variable, in cubic meters. Default: \code{NULL}.
If this argument is \code{NULL}, the defined groups in the dataframe will be used. Default: \code{NULL}.}
}
\value{
A dataframe summarised by the .groups variable(s).
}
\description{
This function can be used to calculate volume with and without bark
of trees in a dataframe.
}
\examples{
library(forestmangr)
data("exfm7")
head(exfm7)

# In order to calculate the volume of each tree, first we
# Calculate the volume by tree section using Smalian's method:
sec_data_vol <- exfm7 \%>\% 
smalianwb("di_wb", "hi", "TREE") \%>\% 
smalianwob("di_wb", "hi", "bark_t", "TREE", bt_mm_to_cm = T)

sec_data_vol

# Now, we summarise the tree's volume:
vol_summarise(sec_data_vol, dbh = "DBH", th = "TH", vwb = "VWB",tree = "TREE", .groups = "STRATA",vwob = "VWOB")

# It's possible to do everything using pipes:
exfm7 \%>\% 
smalianwb("di_wb", "hi", "TREE") \%>\% 
smalianwob("di_wb", "hi", "bark_t", "TREE", bt_mm_to_cm = T) \%>\% 
vol_summarise("DBH", "TH", "VWB", "TREE", "STRATA", "VWOB")

}
\seealso{
Complementary functions:
  \code{\link{smalianwb}}, For calculation of volume with bark using Smalian's method,
  \code{\link{smalianwob}}, For calculation of volume without bark using Smalian's method,
  \code{\link{huberwb}}, for calculation of volume with bark using Huber's method,
  \code{\link{huberwob}}, for calculation of volume without bark Huber's method.
}
\author{
Sollano Rabelo Braga \email{sollanorb@gmail.com}
}
